Created Game/Form/Click events
Created most of the necessary  aspects of the game's display. Created
each of the territory's labels, created click events for each of those
labels, and made the field variables attackingterritory, and
defendingterritory. Created another variable called turnPhase, which
will be used to identify what part of player's turn it is. Inside of the
click events for each territory, I made the territory set
attackingTerritory to it's name.