Changed CheckifAdjacent method to a value return method and began check if transferrable method
CheckifAdjacent method now needs two inputted object variable and will
return if they are adjacent. Check if transferrable uses a while loop to
check if a certain territory is adjacent to another territory and
performs an action. Will be used to eventually see if there is an open
path of player owned territories between two chosen territories in which
to transfer armies between.