Added methods that saves selected Cards
If you click a card, it outlines it. If you click it again the outline
disappears. Also if the card is outlined, its card id (what page its
on/which picture box displays it) is saved to a list. This way if you
change what page of cards your looking at, the cards selected will still
be selected if you want to use them.